Moved discord specific functions into bartleby-discord.py
