+++ bartleby/bartleby.py
@@ -1,10 +1,10 @@
-#import torch
 import queue
 from threading import Thread
 
 import bartleby.configuration as config
 import bartleby.functions.helper_functions as helper_funcs
 import bartleby.functions.IO_functions as io_funcs
+import bartleby.functions.bartleby_discord as discord_funcs
 import bartleby.classes.matrix_class as matrix
 import bartleby.classes.docx_class as docx
 
@@ -65,7 +65,7 @@ def run():
     elif config.MODE == 'discord':
         
         # Start the discord listener
-        discord_listener_thread = Thread(target=io_funcs.discord_listener, args=[
+        discord_listener_thread = Thread(target=discord_funcs.discord_listener, args=[
             config.bot_token,
             docx_instance, 
             users,
