+++ bartleby/functions/helper_functions.py
@@ -37,26 +37,6 @@ def start_logger():
 
     return logger
 
-def start_discord_logger():
-    '''Sets up logging for discord.py'''
-
-    logger = logging.getLogger('discord')
-    logger.setLevel(conf.LOG_LEVEL)
-    logging.getLogger('discord.http').setLevel(logging.INFO)
-
-    handler = RotatingFileHandler(
-        f'{conf.LOG_PATH}/discord.log',
-        encoding='utf-8',
-        maxBytes=32 * 1024 * 1024,  # 32 MiB
-        backupCount=5,  # Rotate through 5 files
-    )
-    dt_fmt = '%Y-%m-%d %H:%M:%S'
-    formatter = logging.Formatter('[{asctime}] [{levelname:<8}] {name}: {message}', dt_fmt, style='{')
-    handler.setFormatter(formatter)
-    logger.addHandler(handler)
-
-    return logger
-
 def check_directory_structure():
     '''Check to make sure we have directories that were not tracked by git'''
